upd: improve get cpu stats

+++ include/top.h
@@ -42,14 +42,14 @@ typedef struct {
 } processes_t;
 
 typedef struct {
-    unsigned long long user;
-    unsigned long long nice;
-    unsigned long long system;
-    unsigned long long idle;
-    unsigned long long iowait;
-    unsigned long long irq;
-    unsigned long long softirq;
-    unsigned long long steal;
+    uint64_t user;
+    uint64_t nice;
+    uint64_t system;
+    uint64_t idle;
+    uint64_t iowait;
+    uint64_t irq;
+    uint64_t softirq;
+    uint64_t steal;
 } cpu_infos_t;
 
 typedef struct {
